# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Testing Shop API",
    "version": "1.0.0",
    "description": "Demo e-shop backend with fault injection for API/UI/Unit testing scenarios."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://{apiHost}",
      "description": "Deployment host",
      "variables": {
        "apiHost": {
          "default": "example.azurecontainerapps.io"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "CartSessionHeader": {
        "in": "header",
        "name": "X-Cart-Session",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Guest cart session key used for cart and checkout operations."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Money": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "currencyCode": {
            "type": "string",
            "example": "CZK"
          }
        },
        "required": [
          "amount",
          "currencyCode"
        ]
      },
...
      "CartLine": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "inStock": {
            "type": "integer"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "lineTotal": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "inStock",
          "price",
          "lineTotal"
        ]
      },
      "Cart": {
        "type": "object",
        "properties": {
          "cartSessionId": {
            "type": "string",
            "format": "uuid"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLine"
            }
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "cartSessionId",
          "items",
          "total"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login as seeded admin or tester user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "tester"
                    ]
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "ADMIN",
                            "TESTER"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "role"
                      ]
                    }
                  },
                  "required": [
                    "token",
                    "user"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cart": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Get cart by session header.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartSessionHeader"
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            },
            "description": "Storefront language used for line names and display money."
          }
        ],
        "responses": {
          "200": {
            "description": "Cart DTO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid session key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cart/items": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add/update/remove a cart line by absolute quantity.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartSessionHeader"
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                },
                "required": [
                  "productId",
                  "quantity"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cart DTO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Validation/service error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}