# backend/src/routes/auth.ts
import { Router } from "express";
import prisma from "../db/prisma";
import { signAdminToken } from "../middleware/adminAuth";
import { UserRole } from "@prisma/client";
import bcrypt from "bcrypt";

const router = Router();

/**
 * @openapi
 * /auth/login:
 *   post:
 *     tags: [Auth]
 *     summary: Login as seeded admin or tester user.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               username:
 *                 type: string
 *                 enum: [admin, tester]
 *               password:
 *                 type: string
 *             required: [username, password]
 *     responses:
 *       200:
 *         description: Login succeeded.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 token:
 *                   type: string
 *                 user:
 *                   type: object
 *                   properties:
 *                     id: { type: integer }
 *                     email: { type: string }
 *                     role:
 *                       type: string
 *                       enum: [ADMIN, TESTER]
 *                   required: [id, email, role]
 *               required: [token, user]
 *       400:
 *         description: Missing credentials.
 *         content:
 *           application/json:
 *             schema: { $ref: '#/components/schemas/ErrorResponse' }
 *       401:
 *         description: Invalid credentials.
 *         content:
 *           application/json:
 *             schema: { $ref: '#/components/schemas/ErrorResponse' }
 */
router.post("/login", async (req, res, next) => {
  try {
    const { username, password } = req.body as {
      username: string;
      password: string;
    };

    if (!username || !password) {
      res.status(400).json({ message: "Username and password are required" });
      return;
    }

    let expectedRole: UserRole | null = null;
    let expectedEmail: string | null = null;

    if (username === "admin") {
      expectedRole = UserRole.ADMIN;
      expectedEmail = "admin@example.com";
    } else if (username === "tester") {
      expectedRole = UserRole.TESTER;
      expectedEmail = "tester@example.com";
    } else {
      res.status(401).json({ message: "Invalid credentials" });
      return;
    }

    const user = await prisma.user.findFirst({
      where: { email: expectedEmail, role: expectedRole },
    });

    if (!user) {
      res.status(500).json({ message: "User is not seeded in the database" });
      return;
    }

    const ok = await bcrypt.compare(password, user.password);
    if (!ok) {
      res.status(401).json({ message: "Invalid credentials" });
      return;
    }

    const token = signAdminToken({ userId: user.id, role: user.role });

    res.json({
      token,
      user: {
        id: user.id,
        email: user.email,
        role: user.role,
      },
    });
  } catch (err) {
    next(err);
  }
});

export default router;